+++ gusto_embedded/src/funcs/jobsAndCompensationsGetCompensation.ts
@@ -18,6 +18,10 @@ import {
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
+import {
+  NotFoundErrorObject,
+  NotFoundErrorObject$inboundSchema,
+} from "../models/errors/notfounderrorobject.js";
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
 import {
@@ -35,7 +39,7 @@ import { Result } from "../types/fp.js";
  * @remarks
  * Compensations contain information on how much is paid out for a job. Jobs may have many compensations, but only one that is active. The current compensation is the one with the most recent `effective_date`.
  *
- * scope: `jobs:read`
+ * scope: `compensations:read`
  */
 export function jobsAndCompensationsGetCompensation(
   client: GustoEmbeddedCore,
@@ -44,6 +48,7 @@ export function jobsAndCompensationsGetCompensation(
 ): APIPromise<
   Result<
     GetV1CompensationsCompensationIdResponse,
+    | NotFoundErrorObject
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -69,6 +74,7 @@ async function $do(
   [
     Result<
       GetV1CompensationsCompensationIdResponse,
+      | NotFoundErrorObject
       | GustoEmbeddedError
       | ResponseValidationError
       | ConnectionError
@@ -164,6 +170,7 @@ async function $do(
 
   const [result] = await M.match<
     GetV1CompensationsCompensationIdResponse,
+    | NotFoundErrorObject
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -176,7 +183,8 @@ async function $do(
     M.json(200, GetV1CompensationsCompensationIdResponse$inboundSchema, {
       key: "Compensation",
     }),
-    M.fail([404, "4XX"]),
+    M.jsonErr(404, NotFoundErrorObject$inboundSchema),
+    M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });
   if (!result.ok) {