+++ gusto_embedded/src/funcs/reimbursementsGetV1RecurringReimbursements.ts
@@ -180,7 +180,7 @@ async function $do(
     | SDKValidationError
   >(
     M.json(200, GetV1RecurringReimbursementsResponse$inboundSchema, {
-      key: "object",
+      key: "Recurring-Reimbursement",
     }),
     M.jsonErr(404, NotFoundErrorObject$inboundSchema),
     M.fail("4XX"),