+++ gusto_embedded/src/react-query/signatoriesInvite.ts
@@ -19,6 +19,7 @@ import {
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
+import { NotFoundErrorObject } from "../models/errors/notfounderrorobject.js";
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
 import { UnprocessableEntityErrorObject } from "../models/errors/unprocessableentityerrorobject.js";
@@ -39,6 +40,7 @@ export type SignatoriesInviteMutationData =
   PostV1CompaniesCompanyUuidSignatoriesInviteResponse;
 
 export type SignatoriesInviteMutationError =
+  | NotFoundErrorObject
   | UnprocessableEntityErrorObject
   | GustoEmbeddedError
   | ResponseValidationError
@@ -53,7 +55,12 @@ export type SignatoriesInviteMutationError =
  * Invite a signatory
  *
  * @remarks
- * Create a signatory with minimal information. This signatory can be invited to provide more information through the `PUT /v1/companies/{company_uuid}/signatories/{signatory_uuid}` endpoint. This will start the identity verification process and allow the signatory to be verified to sign documents.
+ * Creates a signatory with minimal information. This signatory can be invited to provide more information through the [Update a signatory](ref:put-v1-companies-company_uuid-signatories-signatory_uuid) endpoint. This will start the identity verification process and allow the signatory to be verified to sign documents.
+ *
+ * ## Related guides
+ * - [Signatory Events](doc:signatory-events)
+ *
+ * scope: `signatories:manage`
  */
 export function useSignatoriesInviteMutation(
   options?: MutationHookOptions<